move nested public class outside
