+++ oauth2_http/javatests/com/google/auth/oauth2/IdentityPoolCredentialsTest.java
@@ -40,7 +40,6 @@ import com.google.api.client.json.GenericJson;
 import com.google.api.client.util.Clock;
 import com.google.auth.TestUtils;
 import com.google.auth.http.HttpTransportFactory;
-import com.google.auth.oauth2.IdentityPoolCredentials.IdentityPoolCredentialSource;
 import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.IOException;
